device detection: only use v4l2 sources
on systems with libcamera installed, the v4l2 sources are hidden. the
libcamera device info isn't usable for anything at all. including all
devices causes the v4l2 sources to appear.

There used to be code to handle extracting the relevant device info from
pipewire sources, but this would have caused the same device to be
detected twice, so remove that code.